+++ flow/__init__.py
@@ -1,2 +1,2 @@
 APP_NAME = "FlowMobile"
-APP_VERSION = "7.3.8"
+APP_VERSION = "7.4.0"